Fix a bunch of serious warnings.

+++ test/test-directory-visit.c
@@ -66,12 +66,12 @@ directory_visit_callback (const gchar *rel_path,
 	printf ("directory_visit_callback -- rel_path `%s' data `%s'\n",
 		rel_path, (gchar *) data);
 
-	printf ("  File `%s'%s (%s, %s), size %d, mode %04o\n",
+	printf ("  File `%s'%s (%s, %s), size %ld, mode %04o\n",
 		info->name,
 		GNOME_VFS_FILE_INFO_SYMLINK (info) ? " [link]" : "",
 		type_to_string (info->type),
 		gnome_vfs_file_info_get_mime_type (info),
-		info->size, info->permissions);
+		(glong) info->size, info->permissions);
 
 	if (info->name[0] != '.'
 	    || (info->name[1] != '.' && info->name[1] != 0)
